refactor: split up user response and mapped mocks
