+++ projects/client/src/lib/features/auth/queries/currentUserSettingsQuery.spec.ts
@@ -1,7 +1,7 @@
-import { ExtendedUserMappedMock } from '$mocks/data/users/ExtendedUserSettingsMappedMock.ts';
 import { runQuery } from '$test/beds/query/runQuery.ts';
 import { createQuery } from '@tanstack/svelte-query';
 import { describe, expect, it } from 'vitest';
+import { ExtendedUserMappedMock } from '../../../../mocks/data/users/mapped/ExtendedUserSettingsMappedMock.ts';
 import { currentUserSettingsQuery } from './currentUserSettingsQuery.ts';
 
 describe('currentUserSettingsQuery', () => {
