+++ projects/client/src/lib/features/spoilers/_internal/useSpoilerAction.spec.ts
@@ -1,7 +1,7 @@
 import { clone } from '$lib/utils/object/clone';
 import { deepAssign } from '$lib/utils/object/deepAssign';
-import { ExtendedUsersResponseMock } from '$mocks/data/users/ExtendedUserSettingsResponseMock';
-import { WatchedMoviesResponseMock } from '$mocks/data/users/WatchedMoviesResponseMock';
+import { ExtendedUsersResponseMock } from '$mocks/data/users/response/ExtendedUserSettingsResponseMock';
+import { WatchedMoviesResponseMock } from '$mocks/data/users/response/WatchedMoviesResponseMock';
 import { server } from '$mocks/server';
 import { renderStore, setAuthorization } from '$test/beds/store/renderStore';
 import { waitFor } from '@testing-library/svelte';
