+++ projects/client/src/mocks/data/users/mapped/ExtendedUserSettingsMappedMock.ts
