+++ projects/client/src/mocks/data/users/mapped/UserFavoritedMappedMock.ts
