+++ projects/client/src/mocks/data/users/response/ExtendedUserSettingsResponseMock.ts
